LAB1/consumers/ml_classifier/requirements.txt
kafka-python==2.3.0
scikit-learn==1.7.2
pandas==2.2.2
xgboost==3.2.0